\name{netlab}
\docType{data}
\alias{netlab}
\title{Bottom-Up Network Design}
\description{
labels representing sub-networks for the surface air temperature observed by a network of 939 weather stations in 1967.
}
\usage{
data(netlab)
}
\format{
A vector of labels representing sub-networks
}
\source{Oh, H-S. and Li, T-H. (2004) Estimation of global temperature fields from scattered observations by 
a spherical-wavelet-based spatially adaptive method. \emph{Journal of the Royal Statistical Society
Ser.} B, \bold{66}, 221--238.
}

\keyword{datasets}
