\name{SpherWave-internal}
\alias{SpherWave-internal}
\alias{distboun.comp}
\alias{multiselc.comp}
\alias{secdistboun.comp}
\alias{modterritory}
\alias{newgotterritory}
\alias{newterritory}
\title{Internal SpherWave Functions}
\description{
  The \code{distboun.comp}, \code{multiselc.comp}, \code{secdistboun.comp}, \code{modterritory}, \cr
\code{newgotterritory} and \code{newterritory} are internal SpherWave functions.
}
\details{
  These are not to be called by the user.
}
\keyword{internal}
