\name{SpherWave}
\alias{SpherWave-package}
\alias{SpherWave}
\docType{package}
\title{Spherical Wavelets and SW-based spatially adaptive methods}
\description{
   This package carries out
spherical wavelet transform developed Li (1999) and Oh (1999), and
implements wavelet thresholding approaches proposed by Oh and Li
(2004). 
}
\keyword{nonparametric}
